initialized celebration gif frames and created method for displaying winning gif
