+++ pygame_screen.py
@@ -43,18 +43,6 @@ class PyGameScreen:
         self.game_over_frames = [self.images[f"gif_frame{i+1}"] for i in
                                  range(30)]  # Adjust the range based on your GIF frames
 
-        self.frames = {}
-        frame_folder = "seinfeld_gif"  # Replace with your folder name
-        for filename in os.listdir(frame_folder):
-            if filename.endswith((".png", ".jpg", ".jpeg", ".gif")):
-                frame_path = os.path.join(frame_folder, filename)
-                frame_name = os.path.splitext(filename)[0]
-                self.frames[frame_name] = pygame.image.load(frame_path).convert_alpha()
-
-        # Load game over frames
-        self.celebration_frames = [self.frames[f"seinfeld_frame{i+1}"] for i in
-                                   range(30)]  # Adjust the range based on your GIF frames
-
         # Initialize Outside and Inside bets handlers
         self.outside = Outside(self)
         self.inside = Inside(self)
@@ -223,34 +211,6 @@ class PyGameScreen:
                 running = False
 
         pygame.time.wait(0)  # Wait for 0 seconds after animation ends
-
-    def display_winning_gif(self):
-        """Display a GIF animation on the winning screen."""
-        self.screen.fill(self.background_color)
-
-        frame_duration = 100  # Milliseconds between frames
-        start_time = pygame.time.get_ticks()
-
-        running = True
-        while running:
-            current_time = pygame.time.get_ticks()
-            frame_index = (current_time - start_time) // frame_duration % len(self.celebration_frames)
-
-            self.screen.blit(self.celebration_frames[frame_index], (200, 400))  # Adjust position as needed
-
-            pygame.display.flip()
-
-            for event in pygame.event.get():
-                if event.type == pygame.QUIT:
-                    running = False
-                elif event.type == pygame.KEYDOWN or event.type == pygame.MOUSEBUTTONDOWN:
-                    running = False
-
-            if current_time - start_time > 2900:  # Display for 3 seconds
-                running = False
-
-        pygame.time.wait(0)
-
 
     def main_loop(self):
         """Main game loop to handle game logic and user interactions."""
